Code and Asset Cleanup Renamed some texture files and generalised some classes
