+++ src/main/java/com/drevery/scpdisturbance/block/SCP058Egg.java
@@ -1,7 +1,7 @@
 package com.drevery.scpdisturbance.block;
 
 import com.drevery.scpdisturbance.block.base.BaseHorizontalBlock;
-import com.drevery.scpdisturbance.entity.scp058.Scp058TentacleEntity;
+import com.drevery.scpdisturbance.entity.scp.SCP058_TentacleEntity;
 import com.drevery.scpdisturbance.registration.ModEntityTypes;
 import com.drevery.scpdisturbance.utils.Utils;
 import io.github.connortron110.connorsapi.events.blocktickschedule.BlockScheduleHandler;
@@ -9,8 +9,6 @@ import io.github.connortron110.connorsapi.events.blocktickschedule.IScheduledBlo
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.particles.ParticleTypes;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.BlockPos;
@@ -23,7 +21,6 @@ import net.minecraft.world.World;
 import net.minecraftforge.common.util.Constants;
 
 import javax.annotation.Nullable;
-import java.util.stream.Stream;
 
 public class SCP058Egg extends BaseHorizontalBlock implements IScheduledBlockEvent {
     private static final VoxelShape[] SHAPE = Utils.makeHorizontalShapes(VoxelShapes.join(Block.box(
@@ -50,7 +47,7 @@ public class SCP058Egg extends BaseHorizontalBlock implements IScheduledBlockEve
         level.setBlock(pos, Blocks.AIR.defaultBlockState(), Constants.BlockFlags.BLOCK_UPDATE);
         level.playSound(null, pos, SoundEvents.TURTLE_EGG_HATCH, SoundCategory.BLOCKS, 0.7F, 0.9F +  0.2F);
        // level.addParticle(ParticleTypes.ENCHANT, pos.getX() + 0.5, pos.getY(), pos.getZ() + 0.5, 1, 1, 1); // Not working
-        Scp058TentacleEntity entity = ModEntityTypes.SCP_058_TENTACLE.get().create(level);
+        SCP058_TentacleEntity entity = ModEntityTypes.SCP_058_TENTACLE.get().create(level);
         entity.setPos(pos.getX() + 0.5, pos.getY(), pos.getZ() + 0.5);
         level.addFreshEntity(entity);
     }
