+++ src/main/java/com/drevery/scpdisturbance/client/model/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Remove this file when added other models
+ */
+package com.drevery.scpdisturbance.client.model;
