--- src/main/java/com/drevery/scpdisturbance/client/render/scp529/JosieRenderer.java
@@ -1,22 +1,0 @@
-package com.drevery.scpdisturbance.client.render.scp529;
-
-import com.drevery.scpdisturbance.client.model.scp529.JosieModel;
-import com.drevery.scpdisturbance.entity.scp529.JosieEntity;
-import com.drevery.scpdisturbance.utils.Utils;
-import net.minecraft.client.renderer.entity.EntityRendererManager;
-import net.minecraft.client.renderer.entity.MobRenderer;
-import net.minecraft.entity.passive.CatEntity;
-import net.minecraft.util.ResourceLocation;
-
-public class JosieRenderer extends MobRenderer<JosieEntity, JosieModel<JosieEntity>> {
-    protected static final ResourceLocation TEXTURE = Utils.rl("textures/entity/josie.png");
-
-    public JosieRenderer(EntityRendererManager rendererManagerIn) {
-        super(rendererManagerIn, new JosieModel<>(), 0.7F);
-    }
-
-    @Override
-    public ResourceLocation getTextureLocation(JosieEntity entity){
-        return TEXTURE;
-    }
-}
