+++ src/main/java/com/drevery/scpdisturbance/entity/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Remove this file when added other entities
+ */
+package com.drevery.scpdisturbance.entity;
