+++ src/main/java/com/drevery/scpdisturbance/entity/scp/SCP007Entity.java
@@ -1,26 +1,21 @@
-package com.drevery.scpdisturbance.entity.scp049;
+package com.drevery.scpdisturbance.entity.scp;
 
-import com.drevery.scpdisturbance.entity.scp058.Scp058TentacleEntity;
-import com.drevery.scpdisturbance.registration.ModEntityTypes;
 import net.minecraft.block.BlockState;
+import net.minecraft.entity.CreatureEntity;
 import net.minecraft.entity.EntityType;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.ai.goal.*;
-import net.minecraft.entity.monster.MonsterEntity;
-import net.minecraft.entity.monster.ZombieEntity;
+import net.minecraft.entity.ai.goal.LookAtGoal;
+import net.minecraft.entity.ai.goal.LookRandomlyGoal;
+import net.minecraft.entity.ai.goal.WaterAvoidingRandomWalkingGoal;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
 
-import javax.annotation.Nullable;
+public class SCP007Entity extends CreatureEntity {
 
-public class Scp049_2Entity extends MonsterEntity {
-
-    public Scp049_2Entity(EntityType<? extends MonsterEntity> type, World worldIn) {
+    public SCP007Entity(EntityType<? extends CreatureEntity> type, World worldIn) {
         super(type, worldIn);
     }
 
@@ -28,10 +23,8 @@ public class Scp049_2Entity extends MonsterEntity {
         this.goalSelector.addGoal(1, new LookRandomlyGoal(this));
         this.goalSelector.addGoal(1, new LookAtGoal(this, PlayerEntity.class, 10.0F));
         this.goalSelector.addGoal(2, new WaterAvoidingRandomWalkingGoal(this, 0.8D));
-        this.goalSelector.addGoal(2, new MeleeAttackGoal(this, 20.0D, false));
-        this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
-    }
 
+    }
 
     @Override
     protected SoundEvent getAmbientSound() {
@@ -50,6 +43,6 @@ public class Scp049_2Entity extends MonsterEntity {
 
     @Override
     protected void playStepSound(BlockPos pPos, BlockState pBlock) {
-        this.playSound(SoundEvents.METAL_STEP, 0.15F, 1.0F);
+        this.playSound(SoundEvents.ZOMBIE_STEP, 0.15F, 1.0F);
     }
 }
