+++ src/main/java/com/drevery/scpdisturbance/entity/scp/SCP049_CuredEntity.java
@@ -1,21 +1,19 @@
-package com.drevery.scpdisturbance.entity.scp049j;
+package com.drevery.scpdisturbance.entity.scp;
 
+import net.minecraft.block.BlockState;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ai.goal.*;
 import net.minecraft.entity.monster.MonsterEntity;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.block.BlockState;
-import net.minecraft.entity.*;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
+public class SCP049_CuredEntity extends MonsterEntity {
 
-public class Scp049JEntity extends MonsterEntity {
-
-    public Scp049JEntity(EntityType<? extends MonsterEntity> type, World worldIn) {
+    public SCP049_CuredEntity(EntityType<? extends MonsterEntity> type, World worldIn) {
         super(type, worldIn);
     }
 
@@ -23,10 +21,9 @@ public class Scp049JEntity extends MonsterEntity {
         this.goalSelector.addGoal(1, new LookRandomlyGoal(this));
         this.goalSelector.addGoal(1, new LookAtGoal(this, PlayerEntity.class, 10.0F));
         this.goalSelector.addGoal(2, new WaterAvoidingRandomWalkingGoal(this, 0.8D));
-        this.goalSelector.addGoal(2, new MeleeAttackGoal(this, 5.0D, false));
+        this.goalSelector.addGoal(2, new MeleeAttackGoal(this, 20.0D, false));
         this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
     }
-
 
     @Override
     protected SoundEvent getAmbientSound() {
