+++ src/main/java/com/drevery/scpdisturbance/entity/scp/SCP049_JEntity.java
@@ -1,12 +1,10 @@
-package com.drevery.scpdisturbance.entity.scp007;
+package com.drevery.scpdisturbance.entity.scp;
 
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.block.BlockState;
-import net.minecraft.entity.*;
-import net.minecraft.entity.ai.goal.LookAtGoal;
-import net.minecraft.entity.ai.goal.LookRandomlyGoal;
-import net.minecraft.entity.ai.goal.WaterAvoidingRandomWalkingGoal;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.ai.goal.*;
+import net.minecraft.entity.monster.MonsterEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.SoundEvents;
@@ -14,9 +12,9 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 
-public class Scp007Entity extends CreatureEntity {
+public class SCP049_JEntity extends MonsterEntity {
 
-    public Scp007Entity(EntityType<? extends CreatureEntity> type, World worldIn) {
+    public SCP049_JEntity(EntityType<? extends MonsterEntity> type, World worldIn) {
         super(type, worldIn);
     }
 
@@ -24,7 +22,8 @@ public class Scp007Entity extends CreatureEntity {
         this.goalSelector.addGoal(1, new LookRandomlyGoal(this));
         this.goalSelector.addGoal(1, new LookAtGoal(this, PlayerEntity.class, 10.0F));
         this.goalSelector.addGoal(2, new WaterAvoidingRandomWalkingGoal(this, 0.8D));
-
+        this.goalSelector.addGoal(2, new MeleeAttackGoal(this, 5.0D, false));
+        this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, PlayerEntity.class, true));
     }
 
 
@@ -45,6 +44,6 @@ public class Scp007Entity extends CreatureEntity {
 
     @Override
     protected void playStepSound(BlockPos pPos, BlockState pBlock) {
-        this.playSound(SoundEvents.ZOMBIE_STEP, 0.15F, 1.0F);
+        this.playSound(SoundEvents.METAL_STEP, 0.15F, 1.0F);
     }
 }
