+++ src/main/java/com/drevery/scpdisturbance/entity/scp/SCP058_TentacleEntity.java
@@ -1,4 +1,4 @@
-package com.drevery.scpdisturbance.entity.scp058;
+package com.drevery.scpdisturbance.entity.scp;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.Entity;
@@ -14,12 +14,11 @@ import net.minecraft.potion.Effects;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.Difficulty;
 import net.minecraft.world.World;
 
-public class Scp058TentacleEntity extends MonsterEntity {
+public class SCP058_TentacleEntity extends MonsterEntity {
 
-    public Scp058TentacleEntity(EntityType<? extends MonsterEntity> type, World worldIn) {
+    public SCP058_TentacleEntity(EntityType<? extends MonsterEntity> type, World worldIn) {
         super(type, worldIn);
     }
 
@@ -48,16 +47,13 @@ public class Scp058TentacleEntity extends MonsterEntity {
     @Override
     protected void playStepSound(BlockPos pPos, BlockState pBlock) {
     }
-    public boolean doHurtTarget(Entity pEntity) {
-        if (super.doHurtTarget(pEntity)) {
-            if (pEntity instanceof LivingEntity) {
-                    ((LivingEntity)pEntity).addEffect(new EffectInstance(Effects.WITHER,  20, 2));
-                }
 
-            return true;
-        } else {
-            return false;
+    public boolean doHurtTarget(Entity pEntity) {
+        boolean flag = super.doHurtTarget(pEntity);
+        if (flag && pEntity instanceof LivingEntity) {
+            ((LivingEntity)pEntity).addEffect(new EffectInstance(Effects.WITHER,  20, 2));
         }
+        return flag;
     }
 }
