+++ src/main/resources/assets/scpdisturbance/textures/entity/scp007.png
