+++ src/main/resources/assets/scpdisturbance/textures/entity/scp049.png
