+++ src/main/resources/assets/scpdisturbance/textures/entity/scp058-tentacle.png
