+++ src/main/resources/assets/scpdisturbance/textures/entity/scp058.png
