+++ src/main/resources/assets/scpdisturbance/textures/entity/scp529.png
